+++ source/renderer/app/themes/daedalus/cardano.js
@@ -334,6 +334,9 @@ export const CARDANO_THEME_CONFIG = {
   nodeUpdate: {
     '--theme-node-update-background-color': '#efefef',
     '--theme-node-update-title-color': '#5e6066',
+    '--theme-node-sync-icon-color': '#fff',
+    '--theme-node-sync-info-message-background-color': '#fff',
+    '--theme-node-sync-info-message-text-color': '#5e6066',
   },
   notification: {
     '--theme-notification-message-background-color': 'rgba(44, 187, 105, 0.95)',
@@ -639,7 +642,7 @@ export const CARDANO_THEME_CONFIG = {
     '--theme-choice-tabs-bottom-border-color-active': '#5e6066',
   },
   testEnvironment: {
-    '--theme-test-environment-label-background-color': '#ab1700',
+    '--theme-test-environment-label-background-color': '#00ab66',
     '--theme-test-environment-label-text-color': '#fafbfc',
   },
   topBar: {