+++ source/renderer/app/themes/daedalus/dark-blue.js
@@ -337,6 +337,9 @@ export const DARK_BLUE_THEME_CONFIG = {
   nodeUpdate: {
     '--theme-node-update-background-color': '#536370',
     '--theme-node-update-title-color': '#e9f4fe',
+    '--theme-node-sync-info-message-background-color': '#536370',
+    '--theme-node-sync-info-message-text-color': '#e9f4fe',
+    '--theme-node-sync-icon-color': '#fff',
   },
   notification: {
     '--theme-notification-message-background-color': 'rgba(83, 99, 112, 0.95)',
@@ -644,7 +647,7 @@ export const DARK_BLUE_THEME_CONFIG = {
     '--theme-choice-tabs-bottom-border-color-active': '#e9f4fe',
   },
   testEnvironment: {
-    '--theme-test-environment-label-background-color': '#ab1700',
+    '--theme-test-environment-label-background-color': '#00ab66',
     '--theme-test-environment-label-text-color': '#fafbfc',
   },
   topBar: {